Move 'sss_python.*' under 'src/python'
as it's not used outside.
