+++ src/python/pysss_nss_idmap.c
@@ -23,8 +23,9 @@
 
 #include <Python.h>
 
-#include "util/sss_python.h"
 #include "sss_client/idmap/sss_nss_idmap.h"
+
+#include "sss_python.h"
 
 #define SSS_NAME_KEY "name"
 #define SSS_SID_KEY "sid"
